# reduce allocations when binding string/time args